+++ lib/src/extensions/start_of.dart
@@ -1,5 +1,4 @@
 import 'package:moment_dart/src/exception.dart';
-import 'package:moment_dart/src/extensions/constructor.dart';
 import 'package:moment_dart/src/moment.dart';
 
 extension StartOfUnit on DateTime {
@@ -116,6 +115,11 @@ extension StartOfUnit on DateTime {
     return subtract(Duration(days: delta)).startOfDay();
   }
 
+  /// Returns start of the ISO week (always Monday)
+  ///
+  /// Returned object will have same timezone as [this]
+  DateTime startOfIsoWeek() => startOfLocalWeek(DateTime.monday);
+
   /// Returns start of the month
   ///
   /// Returned object will have same timezone as [this]
@@ -183,6 +187,11 @@ extension StartOfUnitMoment on Moment {
         .startOfLocalWeek(weekStart ?? localization.weekStart)
         .toMoment(localization: setLocalization);
   }
+
+  /// Returns start of the ISO week (always Monday)
+  ///
+  /// Returned object will have same timezone as [this]
+  Moment startOfIsoWeek() => startOfLocalWeek(DateTime.monday);
 
   /// Returns start of the month
   ///